+++ CHANGELOG.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased]
 ### Changed
 - Removed Composite from Format and merged the properties into the Format class
+- Added basic Custom Attribute deserialization support
 
 ## [0.8.0]  -  2018-07-17
 ### Added